requirements.txt
argon2-cffi==16.3.0
Babel==2.5.0
bcrypt==3.1.3
cairocffi==0.8.0
CairoSVG==2.0.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
cssselect==1.0.1
cssselect2==0.2.0
Django==1.11
django-bootstrap3==8.2.3
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-discover-runner==1.0
django-phonenumber-field==1.3.0
django-role-permissions==2.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.5
lxml==3.8.0
mysqlclient==1.3.10
olefile==0.44
phonenumberslite==8.8.0
Pillow==4.1.1
psycopg2==2.7.1
pycparser==2.17
Pyphen==0.9.4
pytz==2017.2
requests==2.16.5
six==1.10.0
tinycss==0.4
tinycss2==0.6.0
urllib3==1.21.1
webencodings==0.5.1

